refactor(ai-agent): move protocol modules and drop stream_event shim
