+++ crates/aionui-ai-agent/src/acp_agent.rs
@@ -1,14 +1,14 @@
-use crate::acp_protocol::AcpProtocol;
 use crate::agent_registry::CatalogSender;
 use crate::capability::cli_process::CliAgentProcess;
 use crate::capability::first_message_injector::{InjectionConfig, inject_first_message_prefix};
 use crate::capability::skill_manager::AcpSkillManager;
 use crate::factory::acp_assembler::AcpSessionParams;
 use crate::manager::acp::{AcpSession, AcpSessionEvent, PermissionRouter, PersistedSessionState};
-use crate::shared_kernel::{ModeId, ModelId, SessionId as DomainSessionId};
-use crate::stream_event::{
+use crate::protocol::acp::AcpProtocol;
+use crate::protocol::events::{
     AgentStreamEvent, AvailableCommandsEventData, FinishEventData, SessionAssignedEventData, StartEventData,
 };
+use crate::shared_kernel::{ModeId, ModelId, SessionId as DomainSessionId};
 use crate::types::{AgentStreamChunk, SendMessageData};
 use agent_client_protocol::schema::{
     AgentCapabilities, AvailableCommand, CancelNotification, ContentBlock, LoadSessionRequest, PromptRequest,
