+++ crates/aionui-ai-agent/src/manager/nanobot.rs
@@ -10,7 +10,7 @@ use tracing::{debug, error, info, warn};
 use aionui_common::CommandSpec;
 
 use crate::capability::cli_process::CliAgentProcess;
-use crate::stream_event::AgentStreamEvent;
+use crate::protocol::events::AgentStreamEvent;
 use crate::types::SendMessageData;
 use std::path::PathBuf;
