+++ crates/aionui-ai-agent/src/manager/remote/mod.rs
@@ -17,7 +17,7 @@ use tokio::sync::{Mutex, RwLock, broadcast};
 use tokio_tungstenite::tungstenite::Message;
 use tracing::{debug, error, info, warn};
 
-use crate::stream_event::AgentStreamEvent;
+use crate::protocol::events::AgentStreamEvent;
 use crate::types::SendMessageData;
 
 /// Internal mutable state for the Remote agent.
