+++ crates/aionui-ai-agent/src/protocol/error.rs
