+++ crates/aionui-ai-agent/src/protocol/mod.rs
@@ -1,2 +1,7 @@
+pub mod acp;
 pub mod cli_detect;
+pub mod error;
 pub mod events;
+
+pub use acp::AcpProtocol;
+pub use error::AcpError;
